refactor(effector): decompose the efference crossing — SIGNAL_BOUNDARY P1a
effector.controller.ts held seven jobs. Two are now their own file:

  effector/policy.enforcement.ts     261  arbiter · verdicts · refusals
  effector/escalation.lifecycle.ts   183  hold · resolve · expire
  effector/types.ts                   80  the shapes they pass
  effector.controller.ts       576 → 241  agency↔host · tick ordering

THE SEAMS WERE VERIFIED BEFORE THE CUT, NOT ASSUMED. A field→method map
showed no method touching both the policy fields (_arbiter,
_pendingRefusals) and the escalation fields (_newEscalations,
_activeEscalations, _pendingResolutions) — except _applyVerdict, which is
the ROUTER: deny goes one way, escalate the other. A router across a seam is
the seam working, not a tangle.

A pure move. No behaviour change, import surface unchanged —
effectorController is still the export, so no caller moved. Same discipline
as the planning.engine split.

Two structural notes. The cycle (policy raises an escalation, an escalation
queues a refusal) is wired with arrow closures in the controller rather than
a shared `this`, so neither class imports the other and resolution defers to
call time. And applyPolicyOutcomes STAYS in the controller: its four-step
tick ordering is load-bearing and neither collaborator owns the tick.

_voiceEscalation is left where it is, deliberately. It belongs to none of
the three, but moving it is not a file-boundary question — it is about who
authors the words. escalationAsk is a string template standing in for a
facet that would say it in the Will's own voice, and swapping that is a
behaviour change, not a move.

Cost stated plainly: 576 → 765 lines across four files, most of the growth
headers and the two deps interfaces. What it buys is that P1 and P2 both add
work to this crossing and now have somewhere to add it.

VERIFIED BY COVERAGE, NOT BY A GREEN SUITE. A pure move makes the suite pass
by construction, so passing proves nothing on its own. Mutations injected
into both new files — escalation TTL zeroed, the allow branch disabled —
failed existing tests in each case (2 of 6 in policy.escalation, 2 of 23 in
policy.arbiter), which is what shows the moved code is exercised where it
now lives.

Suite 1825 passed / 227 files; replay.equivalence green.
